Fixed highlighter concurrency and memory issues, updated connection wizard to include game selection, added familiar window, added handling for ooc text and added general highlighting for incoming damage text.

+++ gui/gamedatacontainer.cpp
@@ -12,14 +12,14 @@ GameDataContainer* GameDataContainer::Instance() {
 GameDataContainer::GameDataContainer(QObject *parent) : QObject(parent) {
     converter = DataConverterService::Instance();
     dataService = DataService::Instance();
+    rxNumber.setPattern("(\\d+)");
 }
 
 QStringList GameDataContainer::extractExp(QString exp, bool brief) {
     QStringList result;
 
-    QRegExp rx("(\\d+)");
-    rx.indexIn(exp, 0);
-    result << rx.cap(1);
+    rxNumber.indexIn(exp, 0);
+    result << rxNumber.cap(1);
 
     int index = exp.indexOf('%') + 1;
     if(brief) {
@@ -34,10 +34,13 @@ void GameDataContainer::setExpField(QString name, QString exp) {
     QWriteLocker locker(&lock);
     /* setting exp for exp window */    
     this->exp.insert(name, converter->addNumericStateToExp(exp));
+
     /* extracting exp values for scripting */
-    QStringList expList = this->extractExp(exp, false);
-    dataService->addExpField(name.toLower().toLocal8Bit().data(),
-        expList.at(0).toLocal8Bit().data(), expList.at(1).toLocal8Bit().data());
+    if(dataService->isLoaded()) {
+        QStringList expList = this->extractExp(exp, false);
+        dataService->addExpField(name.toLower().toLocal8Bit().data(),
+            expList.at(0).toLocal8Bit().data(), expList.at(1).toLocal8Bit().data());
+    }
 }
 
 void GameDataContainer::setExpFieldBrief(QString name, QString exp) {
@@ -46,28 +49,36 @@ void GameDataContainer::setExpFieldBrief(QString name, QString exp) {
     this->exp.insert(name, exp);
 
     /* extracting exp brief values for scripting */
-    QStringList expList = this->extractExp(exp, true);
-    dataService->addExpField(name.toLower().toLocal8Bit().data(),
-        expList.at(0).toLocal8Bit().data(), expList.at(1).toLocal8Bit().data());
+    if(dataService->isLoaded()) {
+        QStringList expList = this->extractExp(exp, true);
+        dataService->addExpField(name.toLower().toLocal8Bit().data(),
+            expList.at(0).toLocal8Bit().data(), expList.at(1).toLocal8Bit().data());
+    }
 }
 
 void GameDataContainer::setContainer(QStringList container) {
+    if(dataService->isLoaded()) {
+        dataService->setContainer(container.join(", ").toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     container.removeFirst();
-    dataService->setContainer(container.join(", ").toLocal8Bit().data());
     this->container = container;
 }
 
 void GameDataContainer::setInventory(QStringList inventory) {
+    if(dataService->isLoaded()) {
+        dataService->setInventory(inventory.join(", ").toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     inventory.removeFirst();
-    dataService->setInventory(inventory.join(", ").toLocal8Bit().data());
     this->inventory = inventory;
 }
 
 void GameDataContainer::removeExpField(QString name) {
+    if(dataService->isLoaded()) {
+        dataService->removeExpField(name.toLower().toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
-    dataService->removeExpField(name.toLower().toLocal8Bit().data());
     exp.remove(name);
 }
 
@@ -85,9 +96,11 @@ QStringList GameDataContainer::getContainer() {
 }
 
 void GameDataContainer::setRoomName(QString name) {
+    if(dataService->isLoaded()) {
+        dataService->setRoomTitle(name.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->roomName = name;
-    dataService->setRoomTitle(name.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRoomName() {
@@ -96,9 +109,11 @@ QString GameDataContainer::getRoomName() {
 }
 
 void GameDataContainer::setRoomDesc(QString desc) {
+    if(dataService->isLoaded()) {
+        dataService->setRoomDescription(desc.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->roomDesc = desc;
-    dataService->setRoomDescription(desc.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRoomDesc() {
@@ -107,9 +122,11 @@ QString GameDataContainer::getRoomDesc() {
 }
 
 void GameDataContainer::setRoomObjs(QString objs) {
+    if(dataService->isLoaded()) {
+        dataService->setRoomObjects(objs.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->roomObjs = objs;
-    dataService->setRoomObjects(objs.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRoomObjs() {
@@ -118,9 +135,11 @@ QString GameDataContainer::getRoomObjs() {
 }
 
 void GameDataContainer::setRoomPlayers(QString players) {
+    if(dataService->isLoaded()) {
+        dataService->setRoomPlayers(players.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->roomPlayers = players;
-    dataService->setRoomPlayers(players.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRoomPlayers() {
@@ -129,9 +148,11 @@ QString GameDataContainer::getRoomPlayers() {
 }
 
 void GameDataContainer::setRoomExits(QString exits) {
+    if(dataService->isLoaded()) {
+        dataService->setRoomExits(exits.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->roomExits = exits;
-    dataService->setRoomExits(exits.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRoomExits() {
@@ -152,7 +173,9 @@ QString GameDataContainer::getRoomExtra() {
 void GameDataContainer::setRight(QString right) {
     QWriteLocker locker(&lock);
     this->wieldRight = right;
-    dataService->setWieldRight(right.toLocal8Bit().data());
+    if(dataService->isLoaded()) {
+        dataService->setWieldRight(right.toLocal8Bit().data());
+    }
 }
 
 QString GameDataContainer::getRight() {
@@ -161,9 +184,11 @@ QString GameDataContainer::getRight() {
 }
 
 void GameDataContainer::setRightNoun(QString rightNoun) {
+    if(dataService->isLoaded()) {
+        dataService->setWieldRightNoun(rightNoun.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->wieldRightNoun = rightNoun;
-    dataService->setWieldRightNoun(rightNoun.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getRightNoun() {
@@ -172,9 +197,11 @@ QString GameDataContainer::getRightNoun() {
 }
 
 void GameDataContainer::setLeft(QString left) {
+    if(dataService->isLoaded()) {
+        dataService->setWieldLeft(left.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->wieldLeft = left;
-    dataService->setWieldLeft(left.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getLeft() {
@@ -183,9 +210,11 @@ QString GameDataContainer::getLeft() {
 }
 
 void GameDataContainer::setLeftNoun(QString leftNoun) {
+    if(dataService->isLoaded()) {
+        dataService->setWieldLeftNoun(leftNoun.toLocal8Bit().data());
+    }
     QWriteLocker locker(&lock);
     this->wieldLeftNoun = leftNoun;
-    dataService->setWieldLeftNoun(leftNoun.toLocal8Bit().data());
 }
 
 QString GameDataContainer::getLeftNoun() {
@@ -194,84 +223,100 @@ QString GameDataContainer::getLeftNoun() {
 }
 
 void GameDataContainer::setStanding(bool standing) {
-    //QWriteLocker locker(&lock);
-    dataService->setStanding(standing);
+    if(dataService->isLoaded()) {
+        dataService->setStanding(standing);
+    }
 }
 
 void GameDataContainer::setSitting(bool sitting) {
-    //QWriteLocker locker(&lock);
-    dataService->setSitting(sitting);
+    if(dataService->isLoaded()) {
+        dataService->setSitting(sitting);
+    }
 }
 
 void GameDataContainer::setKneeling(bool kneeling) {
-    //QWriteLocker locker(&lock);
-    dataService->setKneeling(kneeling);
+    if(dataService->isLoaded()) {
+        dataService->setKneeling(kneeling);
+    }
 }
 
 void GameDataContainer::setProne(bool prone) {
-    //QWriteLocker locker(&lock);
-    dataService->setProne(prone);
+    if(dataService->isLoaded()) {
+        dataService->setProne(prone);
+    }
 }
 
 void GameDataContainer::setStunned(bool stunned) {
-    //QWriteLocker locker(&lock);
-    dataService->setStunned(stunned);
+    if(dataService->isLoaded()) {
+        dataService->setStunned(stunned);
+    }
 }
 
 void GameDataContainer::setBleeding(bool bleeding) {
-    //QWriteLocker locker(&lock);
-    dataService->setBleeding(bleeding);
+    if(dataService->isLoaded()) {
+        dataService->setBleeding(bleeding);
+    }
 }
 
 void GameDataContainer::setHidden(bool hidden) {
-    //QWriteLocker locker(&lock);
-    dataService->setHidden(hidden);
+    if(dataService->isLoaded()) {
+        dataService->setHidden(hidden);
+    }
 }
 
 void GameDataContainer::setInvisible(bool invisible) {
-    //QWriteLocker locker(&lock);
-    dataService->setInvisible(invisible);
+    if(dataService->isLoaded()) {
+        dataService->setInvisible(invisible);
+    }
 }
 
 void GameDataContainer::setWebbed(bool webbed) {
-    //QWriteLocker locker(&lock);
-    dataService->setWebbed(webbed);
+    if(dataService->isLoaded()) {
+        dataService->setWebbed(webbed);
+    }
 }
 
 void GameDataContainer::setJoined(bool joined) {
-    //QWriteLocker locker(&lock);
-    dataService->setJoined(joined);
+    if(dataService->isLoaded()) {
+        dataService->setJoined(joined);
+    }
 }
 
 void GameDataContainer::setDead(bool dead) {
-    //QWriteLocker locker(&lock);
-    dataService->setDead(dead);
+    if(dataService->isLoaded()) {
+        dataService->setDead(dead);
+    }
 }
 
 void GameDataContainer::setHealth(int health) {
-    //QWriteLocker locker(&lock);
-    dataService->setHealth(health);
+    if(dataService->isLoaded()) {
+        dataService->setHealth(health);
+    }
 }
 
 void GameDataContainer::setConcentration(int concentration) {
-    //QWriteLocker locker(&lock);
-    dataService->setConcentration(concentration);
+    if(dataService->isLoaded()) {
+        dataService->setConcentration(concentration);
+    }
 }
 
 void GameDataContainer::setSpirit(int spirit) {
-    //QWriteLocker locker(&lock);
-    dataService->setSpirit(spirit);
+    if(dataService->isLoaded()) {
+        dataService->setSpirit(spirit);
+    }
 }
 
 void GameDataContainer::setFatigue(int fatigue) {
-    //QWriteLocker locker(&lock);
-    dataService->setFatigue(fatigue);
+    if(dataService->isLoaded()) {
+        dataService->setFatigue(fatigue);
+    }
 }
 
 void GameDataContainer::setRt(int rt) {
-    //QWriteLocker locker(&lock);
-    if(rt < 0) {
-        rt = 0;
+    if(dataService->isLoaded()) {
+        if(rt < 0) {
+            rt = 0;
+        }
+        dataService->setRt(rt);
     }
-    dataService->setRt(rt);
 }
